==> scrap_journals/1 - parsing module journals for each editor.ipynb <==
# get all journals for each editor and their link to main pages on library genesis

# imports
import pandas as pd
from bs4 import BeautifulSoup
import time

pd.set_option('display.max_rows', 10)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', 1000)
pd.set_option('display.colheader_justify', 'center')
pd.set_option('display.precision', 3)

# define pages
pages = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z', '0-9', 'other']

# Define of the dataframe
df = pd.DataFrame(columns=['Journal','Link', 'Publisher'])

for page in pages:
    print("Parsing page", page, "...")
    content = open("journals/page" + page + ".txt", "r")
    soup = BeautifulSoup(content, 'html.parser')
    # print(soup.prettify())
    content.close()
    
    # Collecting Ddata
    table = soup.find('table')
    table_body = table.find('tbody')
    # print(table_body.prettify())
    rows = table_body.find_all('tr')
    # print(len(rows))
    for row in rows: 
        columns = row.find_all('td') # Find all data for each column
        if(columns != []):
            journal = columns[0].text
            link = columns[0].find('a').get('href')
            publisher = columns[1].text
            df = df.append({'Journal': journal,  'Link': link, 'Publisher': publisher}, ignore_index=True)

df = df[df.Publisher != ""]
df = df.reset_index()
df = df[['Journal', 'Link', 'Publisher']]
display(df)

df_Wiley = df[df.Publisher.str.contains('Wiley')]
df_Wiley[['Optional', 'Journal_Id']] = df_Wiley['Link'].str.split('journals/', expand=True)
df_Wiley = df_Wiley[["Journal", "Journal_Id", "Publisher", "Link" ]]
df_Wiley["Link_Articles"] = "/scimag/?journal=" + df_Wiley["Journal_Id"] + "&page=1" # example: # https://libgen.rs/scimag/?journal=23&page=1
display(df_Wiley)
df_Wiley.to_csv("JournalsWiley.csv", index=False)

df_Springer = df[df.Publisher.str.contains('Springer')]
df_Springer[['Optional', 'Journal_Id']] = df_Springer['Link'].str.split('journals/', expand=True)
df_Springer = df_Springer[["Journal", "Journal_Id", "Publisher", "Link" ]]
df_Springer["Link_Articles"] = "/scimag/?journal=" + df_Springer["Journal_Id"] + "&page=1" # example: # https://libgen.rs/scimag/?journal=23&page=1
display(df_Springer)
df_Springer.to_csv("JournalsSpringer.csv", index=False)

df_Taylor = df[df.Publisher.str.contains('Taylor')]
df_Taylor[['Optional', 'Journal_Id']] = df_Taylor['Link'].str.split('journals/', expand=True)
df_Taylor = df_Taylor[["Journal", "Journal_Id", "Publisher", "Link" ]]
df_Taylor["Link_Articles"] = "/scimag/?journal=" + df_Taylor["Journal_Id"] + "&page=1" # example: # https://libgen.rs/scimag/?journal=23&page=1
display(df_Taylor)
df_Taylor.to_csv("JournalsTaylor.csv", index=False)


df_Elsevier = df[df.Publisher.str.contains('Elsevier')]
df_Elsevier[['Optional', 'Journal_Id']] = df_Elsevier['Link'].str.split('journals/', expand=True)
df_Elsevier = df_Elsevier[["Journal", "Journal_Id", "Publisher", "Link" ]]
df_Elsevier["Link_Articles"] = "/scimag/?journal=" + df_Elsevier["Journal_Id"] + "&page=1" # example: # https://libgen.rs/scimag/?journal=23&page=1
display(df_Elsevier)
df_Elsevier.to_csv("JournalsElsevier.csv", index=False)
