==> scrap_journals/2 - Parsing module journals.ipynb <==
# find the number of pages to scrap for each journal and add information in dataframe.
# this method was then abandoned
# this method was analyzing all articles (from every year, every volume, every issue) for every journal
# after a meeting, we decided it wass better to only scrap 1 issue per year

# imports
import pandas as pd
from bs4 import BeautifulSoup
import time

pd.set_option('display.max_rows', 10)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', 1000)
pd.set_option('display.colheader_justify', 'center')
pd.set_option('display.precision', 3)

#  find the number of pages for each journal and add information in dataframe
publishers = ["Wiley", "Springer", "Taylor", "Elsevier"]
publishers = ["Elsevier"]

for publisher in publishers:
    print("Parsing Publisher", publisher, "...")
    df = pd.read_csv("Journals" + publisher + ".csv")
    # display(df)
    number_of_pages = []
    
    journal_ids = df["Journal_Id"].to_numpy()
    # print(journal_ids)
    
    for journal_id in journal_ids:
        # print(publisher.lower() + "/journal" + str(journal_id) + "/articles_1" + ".txt")
        content = open(publisher.lower() + "/journal" + str(journal_id) + "/articles_1" + ".txt", "r")
        soup = BeautifulSoup(content, 'html.parser')
        # print(soup)
        # print(soup.prettify())
        content.close()
        
        # find number of pages
        data = soup.find("span", {"class": "page_selector"})
        # print(data)
        if data == None: # when there is only 1 page
            number_of_pages.append("1")
        else:
            data = data.text.split(" / ")[-1]
            number_of_pages.append(data)
    
    # add to dataframe
    df["Number_Pages"] = pd.Series(number_of_pages)
    display(df)
    df.to_csv("Journals" + publisher + ".csv", index=False)
        

 # example for 1 page: journal 23

print("Parsing page 23...")
content = open("tryout_journal_23.txt", "r")
soup = BeautifulSoup(content, 'html.parser')
# print(soup.prettify())
content.close()
data = soup.find("span", {"class": "page_selector"})
data = data.text.split(" / ")[-1]
print(data)

# analysis of the time it would take

df_Wiley = pd.read_csv("JournalsWiley.csv")
# display(df_Wiley)
df_Springer = pd.read_csv("JournalsSpringer.csv")
# display(df_Springer)
df_Taylor = pd.read_csv("JournalsTaylor.csv")
# display(df_Taylor)
df_Elsevier = pd.read_csv("JournalsElsevier.csv")
# display(df_Elsevier)

total = df_Wiley["Number_Pages"].sum() + df_Springer["Number_Pages"].sum() + df_Taylor["Number_Pages"].sum() + df_Elsevier["Number_Pages"].sum()
print("Number of pages:", total)
print("Time to scrap:", "%.2f" %(total/3600), "hours")
print("Time to scrap:", "%.2f" %(total/(3600*24)), "days")

print("\nAssumption: 1 relevant pdf every 3 articles")
print("This gives", "%.0f" %(total/3), "pdf documents to download")

print("\nIn order to download a pdf, I need to scrap 2 pages")
print("This gives", "%.0f" %((total/3)*2), "pages to scrap")
print("Time to scrap:", "%.2f" %(((total/3)*2)/3600), "hours")
print("Time to scrap:", "%.2f" %(((total/3)*2)/(3600*24)), "days")

print("\nTotal time:", "%.2f" %((total/3600)+((total/3)*2)/3600), "hours")
print("Total time:", "%.2f" %((total/(3600*24))+(((total/3)*2)/(3600*24))), "days")
print("(Doesn't include time for parsing, probably the same amount of time)")
