==> scrap_journals/3 - Parsing find issue links.ipynb <==
# for each journal in each publisher, find links to 1st issues for volumes corresponding to a specific period of time (
# here, range of time is set from 2017 to 2021
# if there is no issue, the dataframe is filled with None value
# save the links in Journals[Publisher].csv

# imports
import pandas as pd
from bs4 import BeautifulSoup
import time

pd.set_option('display.max_rows', 10)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', 1000)
pd.set_option('display.colheader_justify', 'center')
pd.set_option('display.precision', 3)

publishers = ["Wiley", "Springer", "Taylor", "Elsevier"]
publishers = ["Wiley"]

start_year = 2017
end_year = 2021

for publisher in publishers:
    print("Parsing publisher", publisher, "...")
    df = pd.read_csv("Journals" + publisher + ".csv")
    df = df.drop(columns=['links_2017', 'links_2018', 'links_2019', 'links_2020', 'links_2021'])
    display(df)
    journal_ids = df["Journal_Id"].to_numpy()
    print(journal_ids)
    
    links_journals =[] # list of dictionaries
    
    for journal_id in journal_ids:
        print(journal_id)
        # print(publisher.lower() + "/journal" + str(journal_id) + "/main_page" + ".txt")
        content = open(publisher.lower() + "/journal" + str(journal_id) + "/main_page" + ".txt", "r")
        soup = BeautifulSoup(content, 'html.parser')
        content.close()
        
        # get link of 1 issue of every year from start_year to end_year
        data = soup.find("table", {"class": "journal_issues"})
        # print(data.prettify())
        links = data.find_all('a')
        final_links = {}
        for year in range(start_year, end_year + 1):
            final_links[year] = None
        for i in range(len(links)):
            links[i] = links[i].get('href')
            for year in range(start_year, end_year + 1): # check in the good range of years
                if "year=" + str(year) + "&volume" in links[i]: # check it's not the link for the whole year
                    if "issue" in links[i].split('&')[-1]: # check it has an issue
                        if final_links[year] == None:
                            final_links[year] = links[i]
        # add to list
        links_journals.append(final_links)
    
    # add to dataframe
    for year in range(start_year, end_year + 1): 
        list_links = []
        for links in links_journals: # for each dictionary (1 dictionary for each journal)
            list_links.append(links[year]) # append link corresponding to the good year
        df["links_" + str(year)] = pd.Series(list_links)
    
    display(df)
    df.to_csv("Journals" + publisher + ".csv", index=False)
