==> scrap_journals/5 - Parsing find PDF docs Wiley.ipynb <==
# get the VALID links to relevant PDF
# the previous link got in step 4 are not "valid"
# we needed to do some modifications to get links we can download

# imports
import pandas as pd
from bs4 import BeautifulSoup
import time
import numpy as np
import glob

pd.set_option('display.max_rows', 10)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', 1000)
pd.set_option('display.colheader_justify', 'center')
pd.set_option('display.precision', 3)

df = pd.read_csv("JournalsWileyPDFs.csv")
journals = df[["Journal_Id"]].drop_duplicates()
journals = journals["Journal_Id"].to_numpy()
links_PDF = []
count = 0
for journal in journals:
    files = sorted(glob.glob("wiley/journal" + str(journal) + "/PDFs/*.txt"))
    for file in files:
        # print(file)
        content = open(file, "r")
        soup = BeautifulSoup(content, 'html.parser')
        content.close()
        row = soup.find('embed', {"id": "pdf"})
        if row != None:
            links_PDF.append(row.get('src'))
            count += 1
        else:
            links_PDF.append(None)

#print(links_PDF)
print(count)
df["LinksPDF"] = pd.Series(links_PDF)
display(df)
df.to_csv("JournalsWileyPDFs.csv", index = False)
